backend/layers/shared_layer/requirements.txt
# Shared dependencies for Lambda Layer
boto3>=1.28.0
pydantic>=2.0.0
PyJWT>=2.8.0
bcrypt>=4.0.0
python-dateutil>=2.8.0
email-validator>=2.0.0
